refactor: centralize hooks
