+++ .prettierignore
@@ -1,0 +1,2 @@
+package-lock.json
+styles/tokens.css
