+++ components/docs/use-disclosure.mdx
@@ -7,4 +7,4 @@ Use the `useDisclosure` hook to manage toggleable UI like menus, drawers and acc
 - When a control behaves like a toggle button, add `aria-pressed` in the trigger props.
 - Avoid managing open state manually; the hook centralises semantics and keyboard support.
 
-See the [hook documentation](../hooks/useDisclosure.mdx) for a full API reference.
+See the [hook documentation](../../hooks/useDisclosure.mdx) for a full API reference.
