+++ hooks/usePrefersReducedMotion.mdx
@@ -1,0 +1,7 @@
+import { Meta } from "@storybook/blocks";
+
+<Meta title="Hooks/usePrefersReducedMotion" />
+
+# usePrefersReducedMotion
+
+Determines whether the user has enabled the `prefers-reduced-motion` setting.
