+++ reports/inventory.json
@@ -54,7 +54,8 @@
         "components/TrustedBy/TrustedBy.stories.tsx",
         "components/VisuallyHidden/VisuallyHidden.stories.tsx",
         "components/WhatIBring/WhatIBring.stories.tsx",
-        "components/hooks/useDisclosure.stories.tsx"
+        "hooks/useDisclosure.stories.tsx",
+        "hooks/usePrefersReducedMotion.stories.tsx"
     ],
     "tokens": ["tokens/base.json", "tokens/dark.json", "tokens/light.json"],
     "mixins": [
